codeforge/backend/requirements.replit.txt
# Simplified requirements for Replit deployment
# Core dependencies only - Replit will handle many system packages

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# API and validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.25.1
aiofiles==23.2.1

# Task queue (simplified)
celery==5.3.4

# AI/ML (optional - comment out if not using)
openai==1.3.5
langchain==0.0.340

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS
fastapi-cors==0.0.6